Use `import type` instead of `var [types]` for Flow. Updated deps.
