+++ package.json
@@ -41,12 +41,12 @@
     "hapi": "16.0.1",
     "i18next": "4.1.1",
     "i18next-xhr-backend": "1.2.1",
-    "joi": "10.0.5",
+    "joi": "10.0.6",
     "knex": "0.12.6",
     "localforage": "1.4.3",
     "lodash-es": "4.17.2",
     "moment": "2.17.1",
-    "multihashes": "0.3.0",
+    "multihashes": "0.3.1",
     "node-uuid": "1.4.7",
     "objection": "0.6.2",
     "poop": "2.0.2",
@@ -57,23 +57,23 @@
     "sprint-js": "0.1.0",
     "sqlite3": "3.1.8",
     "string": "3.3.3",
-    "superagent": "3.2.1",
+    "superagent": "3.3.0",
     "tweetnacl": "0.14.5",
     "tweetnacl-util": "0.13.5",
     "uws": "0.12.0",
-    "vee-validate": "2.0.0-beta.14",
+    "vee-validate": "2.0.0-beta.18",
     "velocity-animate": "1.4.0",
     "vue": "2.1.6",
     "vue-router": "2.1.1",
     "vue-script2": "2.0.0",
-    "vuex": "2.0.0",
+    "vuex": "2.1.1",
     "ws": "1.1.1"
   },
   "devDependencies": {
-    "babel-core": "6.20.0",
+    "babel-core": "6.21.0",
     "babel-eslint": "7.1.1",
     "babel-plugin-transform-class-properties": "6.19.0",
-    "babel-plugin-transform-flow-strip-types": "6.18.0",
+    "babel-plugin-transform-flow-strip-types": "6.21.0",
     "babel-plugin-transform-inline-environment-variables": "6.8.0",
     "babel-plugin-transform-runtime": "6.15.0",
     "babel-polyfill": "6.20.0",
